Remove NotFoundPage component
